requirements.txt
streamlit==1.32.0
pandas==2.1.0
numpy==1.26.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
arxiv==1.4.7
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.8.0
plotly==5.17.0
scikit-learn==1.3.0
nltk==3.8.1
python-multipart==0.0.6
transformers>=4.25.0
torch>=1.13.0
sentencepiece==0.1.99
tiktoken==0.5.1
streamlit-option-menu>=0.3.6
streamlit-pdf-viewer>=0.9.0
streamlit-aggrid>=0.3.4
pyvis>=0.2.1
networkx>=2.8.7
pymupdf>=1.21.0
python-jose>=3.3.0
passlib>=1.7.4
python-bibtexparser>=2.0.0b4
streamlit-cookies-manager>=0.2.2
streamlit-shortcuts>=0.2.0
streamlit-theme>=0.1.6
wordcloud>=1.8.2.2
umap-learn>=0.5.3
pyzotero>=1.5.0
scholarly>=1.7.9
pytest>=7.2.0
black>=22.12.0
isort>=5.11.4
mypy>=0.991
diskcache>=5.6.3
